Introduction
With the growth in regard of information systems and the number of
companies that increasingly uses them, there will be a need to perform
security audits to the systems.

Normally in an audit, the auditor that
performs the tests will use several methods and tools to perform the
tests, and each of these tools generate a different report in relation
to the other, and at the end of it all the different reports will be
necessary to make a final report to deliver to the company with the
results of all tests, but currently there isn't still a template for
these reports.

As such there is a need to create a platform for
companies that are responsible for performing audits, with the objective
of aggregating the results of these audits, so that this information is
saved for later use, for when it's made a new audit of the same system,
it is possible check the threats and vulnerabilities that were mitigated
so as the responsible organization can do a better management of their
system security.

Thanks to this a need for a platform of this type had
to be researched, and this investigation is a result of this. The
platform aims to help the people responsible for the audits by providing
the means to insert vulnerabilities manually or by using a file from a
vulnerability scanner. This platform also offers to it's users a
vulnerability scoring system that is very well know by most auditors,
and thanks to it, the vulnerabilities that are registered have more
details that can help everyone understand them.